+++ class/AutodetectFilter.php
@@ -29,9 +29,9 @@ class AutodetectFilter{
     private $_detectionPattern = '/\b([a-z0-9_\.+-]+@[\da-z\.-]+\.[a-z]{2,23})\b/i';
 
 
-    public function __construct($settingsUtil, $shorcodeHandler){
+    public function __construct($settingsManager, $shorcodeHandler){
         // get IDs to exclude
-        $eID = $settingsUtil->getOption('email-autodetect-excludeid');
+        $eID = $settingsManager->getOption('email-autodetect-excludeid');
         
         // filter non numeric chars
         $eID = preg_replace('/[^0-9,]/', '', $eID);
@@ -41,6 +41,31 @@ class AutodetectFilter{
         
         // store shortcode handler instance
         $this->_shortcodeHandler = $shorcodeHandler;
+
+          // filter content ?
+        if ($settingsManager->getOption('email-autodetect-content')){
+            add_filter('the_content', array($this, 'filter'), 50, 1);
+        }
+        
+        // filter excerpt ?
+        if ($settingsManager->getOption('email-autodetect-excerpt')){
+            add_filter('get_the_excerpt', array($this, 'filter'), 50, 1);
+        }
+        
+        // filter comment text ?
+        if ($settingsManager->getOption('email-autodetect-comments')){
+            add_filter('get_comment_text', array($this, 'filterNoExclusion'), 50, 1);
+        }
+        
+        // filter comment excerpt ?
+        if ($settingsManager->getOption('email-autodetect-comments-excerpt')){
+            add_filter('get_comment_excerpt', array($this, 'filterNoExclusion'), 50, 1);
+        }
+        
+        // filter widget text ? 
+        if ($settingsManager->getOption('email-autodetect-widget-text')){
+            add_filter('widget_text', array($this, 'filterNoExclusion'), 50, 1);
+        }
     }
     
     // wp the_content/the_excerpt callback