+++ Cryptex.php
@@ -3,14 +3,14 @@
     Plugin Name: Cryptex - E-Mail Address Protection
     Plugin URI: https://andidittrich.de/go/cryptex
     Description: Advanced Graphical E-Mail Obfuscator which provides image based email address protection using wordpress shortcode and integrated encryption/decryption of addresses
-    Version: 6.0-BETA1
+    Version: 6.0
     Author: Andi Dittrich
     Author URI: https://andidittrich.de
     License: MIT X11 License
 
     ----
     The MIT License (X11 License)
-    Copyright (c) 2010-2016 Andi Dittrich <http://andidittrich.de>
+    Copyright (c) 2010-2017 Andi Dittrich <https://andidittrich.de>
     Permission is hereby granted, free of charge, to any personobtaining a copy of this software and associated documentationfiles (the "Software"), to deal in the Software withoutrestriction, including without limitation the rights to use,copy, modify, merge, publish, distribute, sublicense, and/or sellcopies of the Software, and to permit persons to whom theSoftware is furnished to do so, subject to the followingconditions:
     The above copyright notice and this permission notice shall beincluded in all copies or substantial portions of the Software.
     THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIESOF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE ANDNONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHTHOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISINGFROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OROTHER DEALINGS IN THE SOFTWARE.
@@ -23,7 +23,7 @@
     Plugin Bootstrap Operation
 */
 define('CRYPTEX_INIT', true);
-define('CRYPTEX_VERSION', '6.0-BETA1');
+define('CRYPTEX_VERSION', '6.0');
 define('CRYPTEX_PLUGIN_PATH', dirname(__FILE__));
 define('CRYPTEX_PLUGIN_URL', plugins_url('/cryptex/'));
 define('CRYPTEX_DEFAULT_FONT_PATH', CRYPTEX_PLUGIN_PATH.DIRECTORY_SEPARATOR.'fonts'.DIRECTORY_SEPARATOR);