+++ skltn/SettingsViewHelper.php
@@ -16,9 +16,50 @@ class SettingsViewHelper{
         // load all options
         $this->_config = $settingsManager->getOptions();
     }
+
+    // Generates a selectform  based on settings-name
+    public function displaySelect($title, $optionName, $values, $options=array()){
+        
+        // open setting block
+        $this->settingsHeader($optionName, $title);
+
+        // element attributes
+        $attb = array(
+            'name' => 'cryptex-options[' . $optionName . ']',
+            'id'   => 'cryptex-' . $optionName,
+            'class' => (isset($options['cssClass']) ? $options['cssClass'] : '')
+        );
+
+        // wrap into label
+        echo HtmlUtil::generateTag('label', array(
+            'for' => 'cryptex[' . $optionName . ']'
+        ), false);
+
+        // generate tag, escape attributes
+        echo HtmlUtil::generateTag('select', $attb, false);
+
+        // generate option list
+        foreach ($values as $optionValue=>$optionDescription){
+            $selected = ($this->_config[$optionName] == $optionValue) ? 'selected="selected"' : '';
+            echo '<option value="', esc_attr($optionValue), '" '.$selected.'>', esc_html($optionDescription), '</option>';
+        }
+        
+        echo '</select>';
+
+        // add label text ?
+        if (isset($options['label'])){
+            echo esc_html($options['label']);
+        }
+
+        // close label
+        echo '</label>';
+
+        // close setting block
+        $this->settingsFooter($options);
+    }
     
     // Generates a checkbox based on the settings-name
-    public function displayCheckbox($title, $optionName, $description=''){
+    public function displayCheckbox($title, $optionName, $options=array()){
         // open setting block
         $this->settingsHeader($optionName, $title);
 
@@ -32,12 +73,16 @@ class SettingsViewHelper{
         // dummy checkbox (unchecked value)
         echo HtmlUtil::generateTag('input', $attb, true);
 
+        // wrap into label
+        echo HtmlUtil::generateTag('label', array(
+            'for' => 'cryptex[' . $optionName . ']'
+        ), false);
+
         // element attributes
         $attb = array(
             'name' => 'cryptex-options[' . $optionName . ']',
             'id'   => 'cryptex-' . $optionName,
             'type' => 'checkbox',
-            'title' => $description,
             'value' => '1'
         );
 
@@ -48,44 +93,30 @@ class SettingsViewHelper{
 
         // generate tag, escape attributes
         echo HtmlUtil::generateTag('input', $attb, true);
-        
-        // close setting block
-        $this->settingsFooter();
-    }
-        
-    // Generates a selectform  based on settings-name
-    public function displaySelect($title, $optionName, $values){
-        
-        // open setting block
-        $this->settingsHeader($optionName, $title);
 
-        // element attributes
-        $attb = array(
-            'name' => 'cryptex-options[' . $optionName . ']',
-            'id'   => 'cryptex-' . $optionName
-        );
-
-        // generate tag, escape attributes
-        echo HtmlUtil::generateTag('select', $attb, false);
-
-        // generate option list
-        foreach ($values as $key=>$value){
-            $selected = ($this->_config[$optionName] == $value) ? 'selected="selected"' : '';
-            echo '<option value="', esc_attr($value), '" '.$selected.'>', esc_html($key), '</option>';
+        // add label text ?
+        if (isset($options['label'])){
+            echo esc_html($options['label']);
         }
-        
-        echo '</select>';
 
+        // close label
+        echo '</label>';
+        
         // close setting block
-        $this->settingsFooter();
+        $this->settingsFooter($options);
     }
-        
-        
+ 
+
     // Generates a input-form
-    public function displayInput($title, $optionName, $label, $cssClass=''){
+    public function displayInput($title, $optionName, $options = array()){
 
         // open setting block
         $this->settingsHeader($optionName, $title);
+
+        // wrap into label
+        echo HtmlUtil::generateTag('label', array(
+            'for' => 'cryptex[' . $optionName . ']'
+        ), false);
 
         // element attributes
         $attb = array(
@@ -94,24 +125,23 @@ class SettingsViewHelper{
             'type' => 'text',
             'title' => $title,
             'value' => $this->_config[$optionName],
-            'class' => $cssClass
+            'class' => (isset($options['cssClass']) ? $options['cssClass'] : ''),
+            'placeholder' => (isset($options['placeholder']) ? $options['placeholder'] : '')
         );
-
-        // option selected ?
-        if ($this->_config[$optionName]){ 
-            $attb['checked'] = 'checked';
-        }
 
         // generate tag, escape attributes
         echo HtmlUtil::generateTag('input', $attb, true);
 
-        // generate label
-        echo HtmlUtil::generateTag('label', array(
-            'for' => 'cryptex[' . $optionName . ']'  
-        ), true, '$label');
+        // add label text ?
+        if (isset($options['label'])){
+            echo esc_html($options['label']);
+        }
+
+        // close label
+        echo '</label>';
 
         // close setting block
-        $this->settingsFooter();
+        $this->settingsFooter($options);
     }
 
     private function settingsHeader($optionName, $title){
@@ -119,8 +149,27 @@ class SettingsViewHelper{
         echo '<div class="cryptex-setting"><div class="cryptex-setting-title">', esc_html($title), '</div><div class="cryptex-setting-input">';
     }
 
-    private function settingsFooter(){
-        echo '</div></div>';
+    private function settingsFooter($options = array()){
+
+        // show description ?
+        if (isset($options['description']) && !empty($options['description'])){
+            echo '<div class="cryptex-setting-description">';
+            echo esc_html($options['description']);
+
+            // read-more link available ?
+            if (isset($options['readmore'])){
+                echo ' ', HtmlUtil::generateTag('a', array(
+                    'href' => $options['readmore'],
+                    'title' => 'Read More',
+                    'target' => '_blank'
+                ), true, 'Read More');
+            }
+
+            echo '</div>';
+        }
+
+        // close setting block, close input block
+        echo '</div></div>', "\n";
     }
 
 }